requirements.txt
opencv-python
firebase-admin
RPi.GPIO